refactor: Relocate convert_exceptions, remove utils
